+++ lib/string.zig
@@ -210,7 +210,7 @@ fn parseIntoFields(
             if (string[idx_ptr.* - 1] == 'Z')
                 fields.tz_options = .{ .utc_offset = UTCoffset.UTC }
             else
-                fields.tz_options = .{ .utc_offset = try UTCoffset.fromSeconds(utcoffset, "") };
+                fields.tz_options = .{ .utc_offset = try UTCoffset.fromSeconds(utcoffset, "", false) };
         },
         // 'Z', - ambiguous!
         // 'i', - IANA identifier; would require allocator
@@ -569,7 +569,7 @@ pub fn parseISO8601(string: []const u8, idx_ptr: *usize) !Datetime.Fields {
             if (string[idx_ptr.* - 1] == 'Z')
                 fields.tz_options = .{ .utc_offset = UTCoffset.UTC }
             else
-                fields.tz_options = .{ .utc_offset = try UTCoffset.fromSeconds(utcoffset, "") };
+                fields.tz_options = .{ .utc_offset = try UTCoffset.fromSeconds(utcoffset, "", false) };
             break :parsing;
         },
     }